smarter bundling
