+++ .github/workflows/release.yml
@@ -195,6 +195,8 @@ jobs:
         uses: tauri-apps/tauri-action@v0
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          # Increase Node.js heap for Vite production build (OOM on CI)
+          NODE_OPTIONS: '--max-old-space-size=4096'
           # Frontend env vars (inlined by Vite at build time)
           VITE_SERVER_URL: ${{ vars.VITE_SERVER_URL }}
           # Updater signing
